implemented host remove

+++ src/containers/Hosts/Components/HostCard/HostCard.tsx
@@ -2,13 +2,11 @@ import React from "react";
 import DeleteIcon from "img/delete.svg";
 import DisconnectIcon from "img/disconnect.svg";
 import "./styles.css";
-import { useDispatch, useSelector } from "react-redux";
+import { useDispatch } from "react-redux";
 import { showConfirmationModal } from "redux/actions/modals";
-import { IInitialState } from "redux/types/states";
 import { IHostCardProps } from "./def";
 import { removeHost } from "redux/actions/hosts";
 const HostCard: React.FC<IHostCardProps> = ({ host }: IHostCardProps) => {
-  const storage = useSelector((state: IInitialState) => state.storage);
   const dispatch = useDispatch();
   return (
     <div className="host-card">
@@ -30,6 +28,8 @@ const HostCard: React.FC<IHostCardProps> = ({ host }: IHostCardProps) => {
               showConfirmationModal(
                 "are you sure to delete this host?",
                 (result) => {
+                  console.log(result);
+                  if (!result) return;
                   host.delete();
                   dispatch(removeHost(host));
                 }
